Calculadora-1/Calculadora/app/src/main/java/com/example/calculadora/Operaciones.java
package com.example.calculadora;

public class Operaciones {
    public static double sumar(double a, double b) {
        return a + b;
        }
    public static double restar(double a, double b) {
        return a - b;
    }
    public static double multiplicar(double a, double b) {
        return a * b;
    }
    public static double dividir(double a, double b) {
        return a / b;
    }
    public static int factorial(int n) {
        if (n == 0) {
            return 1;
        } else {
            return n * factorial(n - 1);
        }
    }
}

